+++ src/UpgradeScreen.cpp
@@ -50,7 +50,7 @@ void UpgradeScreen::init() {
 
 void UpgradeScreen::handleInput(sf::RenderWindow& window, const sf::Event& event) {
     if (event.type == sf::Event::MouseButtonPressed && event.mouseButton.button == sf::Mouse::Left) {
-        sf::Vector2f mousePos(event.mouseButton.x, event.mouseButton.y);
+        sf::Vector2f mousePos(static_cast<float>(event.mouseButton.x), static_cast<float>(event.mouseButton.y));
         Player* player = m_gameManager->getPlayer();
         
         for (auto& btn : m_buttons) {
@@ -102,7 +102,7 @@ void UpgradeScreen::handleInput(sf::RenderWindow& window, const sf::Event& event
     
     // Hover effects
     if (event.type == sf::Event::MouseMoved) {
-        sf::Vector2f mousePos(event.mouseMove.x, event.mouseMove.y);
+        sf::Vector2f mousePos(static_cast<float>(event.mouseMove.x), static_cast<float>(event.mouseMove.y));
         for (auto& btn : m_buttons) {
             if (btn.rect.getGlobalBounds().contains(mousePos)) {
                 btn.rect.setFillColor(sf::Color(150, 150, 150));
